+++ internal/compose/documentLink.go
@@ -81,10 +81,12 @@ func createImageLink(serviceNode *ast.MappingValueNode) *protocol.DocumentLink {
 		service := stringNode(serviceNode.Value)
 		if service != nil {
 			linkedText, link := extractImageLink(service.Value)
-			return &protocol.DocumentLink{
-				Range:   createRange(service.GetToken(), len(linkedText)),
-				Target:  types.CreateStringPointer(link),
-				Tooltip: types.CreateStringPointer(link),
+			if linkedText != "" {
+				return &protocol.DocumentLink{
+					Range:   createRange(service.GetToken(), len(linkedText)),
+					Target:  types.CreateStringPointer(link),
+					Tooltip: types.CreateStringPointer(link),
+				}
 			}
 		}
 	}
@@ -221,9 +223,12 @@ func DocumentLink(ctx context.Context, documentURI protocol.URI, doc document.Co
 }
 
 func extractImageLink(nodeValue string) (string, string) {
-	if strings.HasPrefix(nodeValue, "ghcr.io/") {
+	if strings.HasPrefix(nodeValue, "ghcr.io") {
 		idx := strings.LastIndex(nodeValue, ":")
 		if idx == -1 {
+			if len(nodeValue) <= 8 {
+				return "", ""
+			}
 			return nodeValue, fmt.Sprintf("https://%v", nodeValue)
 		}
 		return nodeValue[0:idx], fmt.Sprintf("https://%v", nodeValue[0:idx])
@@ -232,14 +237,20 @@ func extractImageLink(nodeValue string) (string, string) {
 	if strings.HasPrefix(nodeValue, "mcr.microsoft.com") {
 		idx := strings.LastIndex(nodeValue, ":")
 		if idx == -1 {
+			if len(nodeValue) <= 18 {
+				return "", ""
+			}
 			return nodeValue, fmt.Sprintf("https://mcr.microsoft.com/artifact/mar/%v", nodeValue[18:])
 		}
 		return nodeValue[0:idx], fmt.Sprintf("https://mcr.microsoft.com/artifact/mar/%v", nodeValue[18:idx])
 	}
 
-	if strings.HasPrefix(nodeValue, "quay.io/") {
+	if strings.HasPrefix(nodeValue, "quay.io") {
 		idx := strings.LastIndex(nodeValue, ":")
 		if idx == -1 {
+			if len(nodeValue) <= 8 {
+				return "", ""
+			}
 			return nodeValue, fmt.Sprintf("https://quay.io/repository/%v", nodeValue[8:])
 		}
 		return nodeValue[0:idx], fmt.Sprintf("https://quay.io/repository/%v", nodeValue[8:idx])